crawler/requirements.txt
beautifulsoup4==4.13.4
pycryptodome==3.23.0
requests==2.32.3
